Move UserDetailsService to security package.
